requirements.txt
tensorflow>=2.5.0
keras>=2.5.0
numpy>=1.19.5
Pillow>=8.2.0
Flask>=2.0.1
scikit-learn>=0.24.2
h5py>=3.1.0
matplotlib>=3.4.2